Adding in the initial things
For this, I added on the initial lines of code that I had from working in Virtual Studio Code.  This includes the images used, my README file, my index html and my styles css. There is still room for improvement in the project so I will be adding more code in soon or tweaking what is already here.